requirements.txt
Django>=5.2.3
mysqlclient
django-cors-headers
djangorestframework
djangorestframework-simplejwt
